+++ run-test-coverage.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+
+dart pub run test_coverage --port 8789 --min-coverage 40 --no-badge --print-test-output --force-collect --force-exit-after-collect --gen-report --open-report
+
+
+